+++ iamf/cli/proto_conversion/proto_to_obu/parameter_block_generator.cc
@@ -491,12 +491,6 @@ absl::Status PopulateCommonFields(
     const ParamDefinition& param_definition,
     GlobalTimingModule& global_timing_module,
     ParameterBlockWithData& parameter_block_with_data) {
-  if (parameter_block_metadata.has_num_subblocks()) {
-    ABSL_LOG(WARNING)
-        << "Ignoring deprecated `num_subblocks` field in Parameter "
-           "Block OBU. Please remove it.";
-  }
-
   // Get the duration from the parameter definition or the OBU itself as
   // applicable.
   const DecodedUleb128 duration = param_definition.param_definition_mode_ == 1